proyecto-esp32-flask/requirements.txt
flask==2.0.1
paho-mqtt==1.6.1
gunicorn==20.1.0  # Servidor para producción